Increased item ID array size. - Increased limit from 32,767 to 65,535. - Still uses same amount of memory with doubled usable size. - Adjusted CARD0_PET from -256 to 256. - Please run both SQL files after applying this update to avoid values becoming corrupt. - Tested and works on clients 2012-04-10 and up thanks to Cydh. - Removed an unused SQL file.

+++ src/map/cashshop.h
@@ -44,7 +44,7 @@ enum CASHSHOP_BUY_RESULT{
 
 
 struct cash_item_data{
-	uint32 nameid;
+	unsigned short nameid;
 	uint32 price;
 };
